Handle failed responses of the fetch service

+++ src/services/fetch.rs
@@ -1,8 +1,8 @@
+use stdweb::Value;
 use html::Context;
 use super::Task;
 
-pub struct FetchHandle {
-}
+pub struct FetchHandle(Option<Value>);
 
 pub enum Method {
     Get,
@@ -22,7 +22,7 @@ pub trait FetchService<MSG> {
     fn fetch<F, IN, OUT>(&mut self, method: Method, url: &str, data: IN, converter: F) -> FetchHandle
     where
         IN: Into<Option<String>>,
-        OUT: From<String>,
+        OUT: From<Result<String, String>>,
         F: Fn(OUT) -> MSG + 'static;
 }
 
@@ -30,34 +30,50 @@ impl<MSG: 'static> FetchService<MSG> for Context<MSG> {
     fn fetch<F, IN, OUT>(&mut self, method: Method, url: &str, data: IN, converter: F) -> FetchHandle
     where
         IN: Into<Option<String>>,
-        OUT: From<String>,
+        OUT: From<Result<String, String>>,
         F: Fn(OUT) -> MSG + 'static
     {
         let mut tx = self.sender();
-        let callback = move |s: String| {
-            let out = OUT::from(s);
+        let callback = move |success: bool, s: String| {
+            let data = if success { Ok(s) } else { Err(s) };
+            let out = OUT::from(data);
             let msg = converter(out);
             tx.send(msg);
         };
         let method = method.to_argument();
         let body = data.into();
-        js! {
+        let handle = js! {
             var data = {
                 method: @{method},
                 body: @{body},
             };
             var request = new Request(@{url}, data);
             var callback = @{callback};
+            var handle = {
+                interrupt: false,
+                callback,
+            };
             fetch(request).then(function(response) {
-                // TODO Do we need to use blob here?
-                return response.text();
+                if (response.ok) {
+                    // Do we need to use blob here?
+                    return response.text();
+                } else {
+                    throw new Error("Network response was not ok.");
+                }
             }).then(function(data) {
-                callback(data);
-                callback.drop();
+                if (handle.interrupted != true) {
+                    callback(true, data);
+                    callback.drop();
+                }
+            }).catch(function(err) {
+                if (handle.interrupted != true) {
+                    callback(false, data);
+                    callback.drop();
+                }
             });
-        }
-        FetchHandle {
-        }
+            return handle;
+        };
+        FetchHandle(Some(handle))
     }
 }
 
@@ -66,6 +82,11 @@ impl Task for FetchHandle {
         // Fetch API doesn't support request cancelling
         // and we should use this workaround with a flag.
         // In fact, request not canceled, but callback won't be called.
-        unimplemented!();
+        let handle = self.0.take().expect("tried to cancel request fetching twice");
+        js! {
+            var handle = @{handle};
+            handle.interrupted = true;
+            handle.callback.drop();
+        }
     }
 }
